+++ example/lib/in_app_webiew_example.screen.dart
@@ -76,7 +76,7 @@ class _InAppWebViewExampleScreenState extends State<InAppWebViewExampleScreen> {
                   BoxDecoration(border: Border.all(color: Colors.blueAccent)),
                   child: InAppWebView(
                     contextMenu: contextMenu,
-                    initialUrl: "https://github.com/flutter",
+                    initialUrl: "https://www.youtube.com/watch?v=oD5RtLhhubg",
                     // initialFile: "assets/index.html",
                     initialHeaders: {},
                     initialOptions: InAppWebViewGroupOptions(